Task 02 - Data load and preprocess Function made
Task 02 - Completed function for preprocessing 1.

Allows for passing specific ticker and dates to gather data for.

Allows for splitting the data into training and test by a specific date, or randomly with a ratio

Allows for dropping or filling NaN values using multiple methods

Allows for scaling of feature columns

Co-Authored-By: SwinZentron <88529708+SwinZentron@users.noreply.github.com>